feature: create server group will clone scaling policy and scheduled action from source server group (if exists)
