+++ clouddriver-tencent/src/main/groovy/com/netflix/spinnaker/clouddriver/tencent/deploy/description/TencentDeployDescription.groovy
@@ -58,6 +58,8 @@ class TencentDeployDescription extends AbstractTencentCredentialsDescription imp
    */
 
   Source source = new Source()
+  boolean copySourceScalingPoliciesAndActions = true
+
 
   @Canonical
   static class Source implements ServerGroupsNameable {
